+++ src/FilterBundle/FilterEvent.php
@@ -17,4 +17,11 @@ class FilterEvent
      * @Event("BestIt\Commercetools\FilterBundle\Event\Request\ProductProjectionSearchRequestEvent")
      */
     const PRODUCTS_REQUEST_POST = 'best_it_commercetools_filter.event.filter.products.request.post';
+
+    /**
+     * Event for collecting and extend terms
+     *
+     * @Event("BestIt\Commercetools\FilterBundle\Event\Facet\TermEvent")
+     */
+    const FACET_TERM_COLLECT = 'best_it_commercetools_filter.event.filter.facets.term_collect';
 }